requirements.txt
-r ./extra_apps/xadmin/requirements.txt

django==4.2.*
djangorestframework>=3.13.1
django-filter>=22.1
requests>=2.28.2

Pillow

# django debug toolbar
django-debug-toolbar>=3.8.1

# log
# loguru

# retry for exception
# add retry func for current requests
# - make_request
# docs reference:
# - https://tenacity.readthedocs.io/en/latest/
# tenacity

# Celery
# celery==5.3.4
# django-celery-beat==2.5.0
# django-celery-results==2.4.0

# https://pypi.org/project/mysqlclient/#files
mysqlclient==2.1.1
pymysql
django-redis==5.2.0

# deploy
# DEBUG model : daphne
# https://docs.gunicorn.org/en/stable/
gunicorn
# https://www.uvicorn.org/deployment/
uvicorn[standard]

# docker cpu count
psutil

# python-dateutil

# web jwt
pyjwt

# wxwork
# wechatpy  version of 1.18 is not latest, so custom development this project(./extra_apps/wechatpy)
# wechatpy
# wechatpy depend on crypto
cryptography
optionaldict

# old env pip
xmltodict
simplejson

# excel
openpyxl

# dotenv
python-dotenv

lxml
